requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pymongo==3.12.3
sqlparse==0.2.4
dnspython==2.4.2
python-dotenv==1.0.0
